# openapi.yaml
openapi: "3.0.3"
info:
  title: EngageIU API
  version: "1.0.0"
  description: >
    QR/code-based campus event attendance leaderboard system for Indiana University.
    Students earn points by entering check-in codes at IU events.
    Built for Luddy Hacks — Dynamic Leaderboard/Ranking System case.
  contact:
    name: EngageIU Team
    email: iuhacks@iu.edu
servers:
  - url: http://localhost:8000
    description: Local development server
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Token:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
...
    Error:
      type: object
      properties:
        detail:
          type: string
          example: "Not found"
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Admin login
      description: Authenticate as admin and receive a JWT bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: "admin"
                password:
                  type: string
                  example: "engageiu2025"
      responses:
        "200":
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        "401":
          description: Incorrect credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...